upgrade to arkworks, add m1 and examples to encode to contract
