+++ bls/bls.h
@@ -175,6 +175,8 @@ bool serialize_public_key_uncompressed(const PublicKey *in_public_key, uint8_t *
 
 bool serialize_signature(const Signature *in_signature, uint8_t **out_bytes, int *out_len);
 
+bool serialize_signature_uncompressed(const Signature *in_signature, uint8_t **out_bytes, int *out_len);
+
 bool sign_message(const PrivateKey *in_private_key,
                   const uint8_t *in_message,
                   int in_message_len,
