+++ bls/bls_darwin64_m1.go
@@ -1,0 +1,9 @@
+// +build darwin,amd64,!ios
+
+package bls
+
+/*
+#cgo LDFLAGS: -L${SRCDIR}/../libs/aarch64-apple-darwin -lbls_snark_sys -ldl -lm
+*/
+import "C"
+
