+++ examples/bls_example/go.mod
@@ -1,0 +1,9 @@
+module github.com/celo-org/celo-bls-go/examples/bls_example
+
+go 1.13
+
+require (
+	github.com/celo-org/celo-bls-go/bls v0.0.0-00010101000000-000000000000
+)
+
+replace github.com/celo-org/celo-bls-go/bls => ../../
