+++ examples/prepare_for_contract/go.mod
@@ -1,0 +1,9 @@
+module github.com/celo-org/celo-bls-go/examples/prepare_for_contract
+
+go 1.13
+
+require (
+	github.com/celo-org/celo-bls-go v0.0.0-00010101000000-000000000000
+)
+
+replace github.com/celo-org/celo-bls-go => ../../
