+++ scripts/release.sh
@@ -10,21 +10,14 @@ fi
 pushd celo-bls-snark-rs/crates/bls-snark-sys
 
 export RUSTFLAGS="-Ccodegen-units=1"
-rustup default 1.47.0
+rustup default 1.52.1
 cargo install cargo-lipo -f
 cargo install cargo-strip -f
 
-#rustup target add i686-unknown-linux-gnu
 rustup target add x86_64-unknown-linux-gnu
-#rustup target add arm-unknown-linux-gnueabi
-#rustup target add arm-unknown-linux-gnueabihf
 rustup target add aarch64-unknown-linux-gnu
-#rustup target add mips-unknown-linux-gnu
-#rustup target add mipsel-unknown-linux-gnu
-#rustup target add mips64-unknown-linux-gnuabi64
-#rustup target add mips64el-unknown-linux-gnuabi64
 rustup target add x86_64-apple-darwin
-#rustup target add i686-pc-windows-gnu
+rustup target add aarch64-apple-darwin
 rustup target add x86_64-pc-windows-gnu
 rustup target add aarch64-linux-android
 rustup target add armv7-linux-androideabi
@@ -41,30 +34,14 @@ cargo build --release --target=i686-linux-android --lib -p bls-snark-sys
 cargo strip --target i686-linux-android
 cargo build --release --target=x86_64-linux-android --lib -p bls-snark-sys
 cargo strip --target x86_64-linux-android
-#cargo build --release -p bls-snark-sys
-#cargo strip
-#cargo build --target=i686-unknown-linux-gnu --release -p bls-snark-sys
-#cargo strip --target i686-unknown-linux-gnu
 cargo build --target=x86_64-unknown-linux-gnu --release -p bls-snark-sys
 cargo strip --target x86_64-unknown-linux-gnu
-#cargo build --target=arm-unknown-linux-gnueabi --release -p bls-snark-sys
-#cargo strip --target arm-unknown-linux-gnueabi
-#cargo build --target=arm-unknown-linux-gnueabihf --release -p bls-snark-sys
-#cargo strip --target arm-unknown-linux-gnueabihf
 cargo build --target=aarch64-unknown-linux-gnu --release -p bls-snark-sys
 cargo strip --target aarch64-unknown-linux-gnu
-#cargo build --target=mips-unknown-linux-gnu --release -p bls-snark-sys
-#cargo strip --target mips-unknown-linux-gnu
-#cargo build --target=mipsel-unknown-linux-gnu --release -p bls-snark-sys
-#cargo strip --target mipsel-unknown-linux-gnu
-#cargo build --target=mips64-unknown-linux-gnuabi64 --release -p bls-snark-sys
-#cargo strip --target mips64-unknown-linux-gnuabi64
-#cargo build --target=mips64el-unknown-linux-gnuabi64 --release -p bls-snark-sys
-#cargo strip --target mips64el-unknown-linux-gnuabi64
 cargo build --target=x86_64-apple-darwin --release -p bls-snark-sys
 cargo strip --target x86_64-apple-darwin
-#cargo build --target=i686-pc-windows-gnu --release -p bls-snark-sys
-#cargo strip --target i686-pc-windows-gnu
+cargo build --target=aarch64-apple-darwin --release -p bls-snark-sys
+cargo strip --target aarch64-apple-darwin
 cargo build --target=x86_64-pc-windows-gnu --release -p bls-snark-sys
 cargo strip --target x86_64-pc-windows-gnu
 cargo build --target=x86_64-unknown-linux-musl --release -p bls-snark-sys
